opyimization: Adjust the copywriting and layout of the monitoring strategy page
