+++ web/src/app/monitor/(pages)/event/strategy/detail/metricDefinitionForm.tsx
@@ -1,65 +1,47 @@
 import React from 'react';
-import {
-  Form,
-  Input,
-  Button,
-  Segmented,
-  Select,
-  Tooltip,
-  InputNumber,
-} from 'antd';
-import { PlusOutlined } from '@ant-design/icons';
+import { Form, Input, Segmented, Select, Tooltip, InputNumber } from 'antd';
 import { useTranslation } from '@/utils/i18n';
 import { SegmentedItem, IndexViewItem, FilterItem } from '@/app/monitor/types';
-import { SourceFeild, StrategyFields } from '@/app/monitor/types/event';
+import { StrategyFields } from '@/app/monitor/types/event';
 import { useScheduleList, useMethodList } from '@/app/monitor/hooks/event';
 import { SCHEDULE_UNIT_MAP } from '@/app/monitor/constants/event';
 import ConditionSelector from './conditionSelector';
-import strategyStyle from '../index.module.scss';
 
 const { Option } = Select;
 const { TextArea } = Input;
 
 interface MetricDefinitionFormProps {
   pluginList: SegmentedItem[];
-  source: SourceFeild;
   metric: string | null;
   metricsLoading: boolean;
   labels: string[];
   conditions: FilterItem[];
   groupBy: string[];
-  unit: string;
   periodUnit: string;
   originMetricData: IndexViewItem[];
   monitorName: string;
   onCollectTypeChange: (id: string) => void;
-  onOpenInstModal: () => void;
   onMetricChange: (val: string) => void;
   onFiltersChange: (filters: FilterItem[]) => void;
   onGroupChange: (val: string[]) => void;
-  onUnitChange: (val: string) => void;
   onPeriodUnitChange: (val: string) => void;
   isTrap: (getFieldValue: any) => boolean;
 }
 
 const MetricDefinitionForm: React.FC<MetricDefinitionFormProps> = ({
   pluginList,
-  source,
   metric,
   metricsLoading,
   labels,
   conditions,
   groupBy,
-  unit,
   periodUnit,
   originMetricData,
   monitorName,
   onCollectTypeChange,
-  onOpenInstModal,
   onMetricChange,
   onFiltersChange,
   onGroupChange,
-  onUnitChange,
   onPeriodUnitChange,
   isTrap,
 }) => {
@@ -69,13 +51,6 @@ const MetricDefinitionForm: React.FC<MetricDefinitionFormProps> = ({
   const SCHEDULE_LIST = useScheduleList();
 
   // 验证函数移到组件内部
-  const validateAssets = async () => {
-    if (!source.values.length) {
-      return Promise.reject(new Error(t('monitor.assetValidate')));
-    }
-    return Promise.resolve();
-  };
-
   const validateMetric = async () => {
     if (!metric) {
       return Promise.reject(new Error(t('monitor.events.metricValidate')));
@@ -94,9 +69,12 @@ const MetricDefinitionForm: React.FC<MetricDefinitionFormProps> = ({
   return (
     <>
       <Form.Item
-        className={strategyStyle.clusterLabel}
         name="collect_type"
-        label={<span className={strategyStyle.label}></span>}
+        label={
+          <span className="w-[100px]">
+            {t('monitor.events.collectionTemplate')}
+          </span>
+        }
         rules={[{ required: true, message: t('common.required') }]}
       >
         <Segmented
@@ -131,55 +109,29 @@ const MetricDefinitionForm: React.FC<MetricDefinitionFormProps> = ({
               />
             </Form.Item>
           ) : (
-            <>
-              <Form.Item<StrategyFields>
-                label={<span className="w-[100px]">{t('monitor.source')}</span>}
-                name="source"
-                rules={[{ required: true, validator: validateAssets }]}
-              >
-                <div>
-                  <div className="flex">
-                    {t('common.select')}
-                    <span className="text-[var(--color-primary)] px-[4px]">
-                      {source.values.length}
-                    </span>
-                    {t('monitor.assets')}
-                    <Button
-                      className="ml-[10px]"
-                      icon={<PlusOutlined />}
-                      size="small"
-                      onClick={onOpenInstModal}
-                    ></Button>
-                  </div>
-                  <div className="text-[var(--color-text-3)] mt-[10px]">
-                    {t('monitor.events.setAssets')}
-                  </div>
-                </div>
-              </Form.Item>
-              <Form.Item<StrategyFields>
-                name="metric"
-                label={<span className="w-[100px]">{t('monitor.metric')}</span>}
-                rules={[{ validator: validateMetric, required: true }]}
-              >
-                <ConditionSelector
-                  data={{
-                    metric,
-                    filters: conditions,
-                    group: groupBy,
-                  }}
-                  metricData={originMetricData}
-                  labels={labels}
-                  loading={metricsLoading}
-                  monitorName={monitorName}
-                  onMetricChange={onMetricChange}
-                  onFiltersChange={onFiltersChange}
-                  onGroupChange={onGroupChange}
-                />
-                <div className="text-[var(--color-text-3)]">
-                  {t('monitor.events.setDimensions')}
-                </div>
-              </Form.Item>
-            </>
+            <Form.Item<StrategyFields>
+              name="metric"
+              label={<span className="w-[100px]">{t('monitor.metric')}</span>}
+              rules={[{ validator: validateMetric, required: true }]}
+            >
+              <ConditionSelector
+                data={{
+                  metric,
+                  filters: conditions,
+                  group: groupBy,
+                }}
+                metricData={originMetricData}
+                labels={labels}
+                loading={metricsLoading}
+                monitorName={monitorName}
+                onMetricChange={onMetricChange}
+                onFiltersChange={onFiltersChange}
+                onGroupChange={onGroupChange}
+              />
+              <div className="text-[var(--color-text-3)]">
+                {t('monitor.events.setDimensions')}
+              </div>
+            </Form.Item>
           )
         }
       </Form.Item>
@@ -194,7 +146,9 @@ const MetricDefinitionForm: React.FC<MetricDefinitionFormProps> = ({
             <Form.Item<StrategyFields>
               required
               label={
-                <span className="w-[100px]">{t('monitor.events.method')}</span>
+                <span className="w-[100px]">
+                  {t('monitor.events.convergenceMethod')}
+                </span>
               }
             >
               <Form.Item
@@ -211,7 +165,7 @@ const MetricDefinitionForm: React.FC<MetricDefinitionFormProps> = ({
                   style={{
                     width: '300px',
                   }}
-                  placeholder={t('monitor.events.method')}
+                  placeholder={t('monitor.events.convergenceMethod')}
                   showSearch
                 >
                   {METHOD_LIST.map((item) => (
@@ -242,45 +196,10 @@ const MetricDefinitionForm: React.FC<MetricDefinitionFormProps> = ({
       <Form.Item<StrategyFields>
         required
         label={
-          <span className="w-[100px]">{t('monitor.events.frequency')}</span>
+          <span className="w-[100px]">
+            {t('monitor.events.convergenceCycle')}
+          </span>
         }
-      >
-        <Form.Item
-          name="schedule"
-          noStyle
-          rules={[
-            {
-              required: true,
-              message: t('common.required'),
-            },
-          ]}
-        >
-          <InputNumber
-            min={SCHEDULE_UNIT_MAP[`${unit}Min`]}
-            max={SCHEDULE_UNIT_MAP[`${unit}Max`]}
-            precision={0}
-            addonAfter={
-              <Select
-                value={unit}
-                style={{ width: 120 }}
-                onChange={onUnitChange}
-              >
-                {SCHEDULE_LIST.map((item) => (
-                  <Option key={item.value} value={item.value}>
-                    {item.label}
-                  </Option>
-                ))}
-              </Select>
-            }
-          />
-        </Form.Item>
-        <div className="text-[var(--color-text-3)] mt-[10px]">
-          {t('monitor.events.setFrequency')}
-        </div>
-      </Form.Item>
-      <Form.Item<StrategyFields>
-        required
-        label={<span className="w-[100px]">{t('monitor.events.period')}</span>}
       >
         <Form.Item
           name="period"
